greenhouse_sharepoint_mapper/requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.28.0
pandas>=1.5.0
msal>=1.24.0
msgraph-core>=0.2.2
